+++ ebpub/ebpub/openblockapi/itemquery.py
@@ -41,6 +41,7 @@ from ebpub.utils.dates import parse_date
 from ebpub.db.models import NewsItem
 from ebpub.streets.models import Place
 import pyrfc3339
+import re
 
 __all__ = ['build_item_query', 'build_place_query']
 
@@ -61,7 +62,9 @@ def build_item_query(request):
     # some different ordering may be more optimal here /
     # some index could be specifically created.
     # Also this could be rewritten to use ebpub.db.schemafilter
-    filters = [_schema_filter, _daterange_filter, _predefined_place_filter,
+    filters = [_schema_filter,
+               _id_filter,
+               _daterange_filter, _predefined_place_filter,
                _radius_filter, _bbox_filter, _attributes_filter, _order_by,
                _object_limit]
 
@@ -82,18 +85,29 @@ def _schema_filter(query, params, state):
 
     slug = params.pop('type', None)
     if slug is not None:
+        if isinstance(slug, basestring):
+            query = query.filter(schema__slug=slug)
+        else:
+            query = query.filter(schema__slug__in=slug)
         state['schema_slug'] = slug
-        query = query.filter(schema__slug=slug)
+    return query, params, state
+
+def _id_filter(query, params, state):
+    """
+    handles filtering items by explicit IDs
+    parameters: 'id'
+    """
+    ids = params.pop('id', None)
+    if ids is not None:
+        if isinstance(ids, basestring):
+            ids = [i for i in re.split(r'[^\d]+', ids) if i.strip()]
+        if ids:
+            query = query.filter(id__in=ids)
     return query, params, state
 
 
 def _attributes_filter(query, params, state):
     # not implemented yet
-    #
-    # schema_slug = state.get('schema_slug')
-    # if schema_slug is None or len(params) == 0: 
-    #     return query, params, state
-        
     return query, params, state
 
 def _daterange_filter(query, params, state):
@@ -132,26 +146,38 @@ def _predefined_place_filter(query, params, state):
     handles filtering by predefined place (newsitemlocation)
     parameters: locationid
     """
-    locationid = params.get('locationid')
+    locationid = params.pop('locationid', None)
     if locationid is None: 
         return query, params, state
-        
-    del params['locationid']
-    
     if state.get('has_geo_filter') == True: 
         raise QueryError('Only one geographic filter may be specified')
 
-    try:
-        loctypeslug, locslug = locationid.split('/')
-        query = query.filter(newsitemlocation__location__slug=locslug,
-                             newsitemlocation__location__location_type__slug=loctypeslug) 
-    except ValueError: 
-        raise QueryError('Invalid location identifier "%s"' % locationid)
+    if isinstance(locationid, basestring):
+        ids = [locationid]
+    else:
+        ids = locationid
+    queries = []
+    from django.db.models import Q
+    # Need to build up an OR query, so we use django Q objects.
+    for loc in ids:
+        try:
+            loctypeslug, locslug = loc.split('/')
+            queries.append(Q(newsitemlocation__location__slug=locslug,
+                             newsitemlocation__location__location_type__slug=loctypeslug))
+        except ValueError: 
+            raise QueryError('Invalid location identifier "%s"' % loc)
+    if queries:
+        loc_query = queries[0]
+        for q in queries[1:]:
+            loc_query = loc_query | q
+
+        # Could get overlaps?
+        query = query.filter(loc_query).distinct()
 
     state['has_geo_filter'] = True
-    
     return query, params, state
-    
+
+
 def _bbox_filter(query, params, state):
     """
     handles filtering by a bounding box region
@@ -293,9 +319,10 @@ def build_place_query(params):
 
 def _copy_nomulti(d):
     """
-    make a copy of django wack-o immutable query mulit-dict
+    make a copy of django wack-o immutable query multi-dict
     making single item values non-lists.
     """
+    d = dict(d) # Work around request.GET.items() not giving all values
     r = {}
     for k,v in d.items():
         try: