+++ ebpub/ebpub/streets/admin.py
@@ -366,8 +366,11 @@ class PlaceAdmin(OSMModelAdmin):
 
 class BlockAdmin(OSMModelAdmin):
 
-    list_display = ('pretty_name', 'street', 'suffix', 'left_zip', 'right_zip', 'left_city', 'right_city')
-    list_filter = ('suffix', 'left_city', 'right_city', 'left_zip', 'right_zip')
+    list_display = ('pretty_name', 'predir', 'street', 'suffix', 'postdir',
+                    'left_zip', 'right_zip', 'left_city', 'right_city')
+    list_filter = ('suffix', 'left_city', 'right_city', 'left_zip', 'right_zip',
+                   'predir', 'postdir',
+                   )
     search_fields = ('pretty_name',)
     readonly_fields = ('from_num', 'to_num',)